Add logic in TaskService.java

+++ src/main/java/ru/leska/taskapp/repository/TaskRepository.java
@@ -1,9 +1,9 @@
 package ru.leska.taskapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.scheduling.config.Task;
 import org.springframework.stereotype.Repository;
+import ru.leska.taskapp.models.Task;
 
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task, Integer> {
 }
